+++ objcToxDemo/OCTCallsViewController.m
@@ -73,7 +73,7 @@
                            @"friend receiving audio: %d\n"
                            @"friend sending video: %d\n"
                            @"friend receiving Video: %d\n",
-                           call.chat.uniqueIdentifier, call.status, call.callDuration, call.friendSendingAudio, call.friendAcceptingAudio,
+                           call.chat.uniqueIdentifier, (long)call.status, call.callDuration, call.friendSendingAudio, call.friendAcceptingAudio,
                            call.friendSendingVideo, call.friendAcceptingVideo];
 
     return cell;
